005-FIX-UPDATE-GOOD

+++ src/main/java/com/dgd/service/UserService.java
@@ -22,6 +22,8 @@ import org.springframework.web.multipart.MultipartFile;
 import javax.servlet.http.Cookie;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import java.util.ArrayList;
+import java.util.List;
 import java.util.Optional;
 import java.util.concurrent.TimeUnit;
 import static com.dgd.exception.message.AuthErrorMessage.*;
@@ -111,15 +113,22 @@ public class UserService {
         User user = userRepository.findById(updateUserDto.getId())
                .orElseThrow(() -> new AuthenticationException(USER_NOT_FOUND));
 
-        if (!DEFAULT.equals(user.getProfileUrl())){
-            s3Service.deleteImage(user);
-        }
+
+
+
         if (userRepository.findByNickName(updateUserDto.getNickName()).isPresent()) {
             throw new AuthenticationException(DUPLICATED_NICKNAME);
         }
+        String userImage = "";
+        if (multipartFile != null) {
+            if (!DEFAULT.equals(user.getProfileUrl())) {
+                s3Service.deleteImage(user);
+            }
+            userImage = s3Service.uploadUserImage(multipartFile);
+        }
 
         user.setLatAndLon(latitude, longitude);
-        user.update(updateUserDto, s3Service.uploadUserImage(multipartFile));
+        user.update(updateUserDto,userImage);
         user.authorizeUser();
 
        return user;
